Validations updated on menu-items-categories

+++ src/menu-item-category/menu-item-category.controller.ts
@@ -7,11 +7,14 @@ import {
   Body,
   Param,
   UsePipes,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 
 import { ValidationPipe } from '../shared/validation.pipe';
 import { MenuItemCategoryService } from './menu-item-category.service';
 import { MenuItemCategoryDTO } from './menu-item-category.dto';
+var validate = require('uuid-validate');
 @Controller('categories')
 export class MenuItemCategoryController {
   constructor(private menuItemCategoryService: MenuItemCategoryService) {}
@@ -28,6 +31,10 @@ export class MenuItemCategoryController {
 
   @Get(':id')
   getCategoryById(@Param('id') id: string) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exceptio
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.menuItemCategoryService.read(id);
   }
 
@@ -37,11 +44,19 @@ export class MenuItemCategoryController {
     @Param('id') id: string,
     @Body() data: Partial<MenuItemCategoryDTO>,
   ) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exception
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.menuItemCategoryService.update(id, data);
   }
 
   @Delete(':id')
   deleteCategory(@Param('id') id: string) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exception
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.menuItemCategoryService.destroy(id);
   }
 }
